final code

+++ app/src/main/java/com/example/weatherapp/data/DailyForecast.kt
@@ -1,4 +1,5 @@
 package com.example.weatherapp.data
+
 import com.google.gson.annotations.SerializedName
 
 
@@ -9,13 +10,13 @@ data class DailyForecast(
     @SerializedName("sunset") val sunset: Int,
     @SerializedName("temp") val temp: Temp,
     @SerializedName("feels_like") val feels_like: FeelsLike,
-    @SerializedName("pressure") val pressure: Int,
-    @SerializedName("humidity") val humidity: Int,
+    @SerializedName("pressure") val pressure: Double,
+    @SerializedName("humidity") val humidity: Double,
     @SerializedName("dew_point") val dew_point: Double,
     @SerializedName("wind_speed") val wind_speed: Double,
     @SerializedName("wind_deg") val wind_deg: Int,
     @SerializedName("weather") val weather: List<Weather>,
     @SerializedName("clouds") val clouds: Int,
-    @SerializedName("pop") val pop: Int,
+    @SerializedName("pop") val pop: Double,
     @SerializedName("uvi") val uvi: Double
 )
